requirements.txt
asgiref==3.7.2
boto3==1.26.145
botocore==1.29.145
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
django-jet-reboot==1.3.3
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
gunicorn==20.1.0
idna==3.4
jmespath==0.10.0
oauthlib==3.2.2
pathspec==0.10.1
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pypiwin32==223
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywin32==306
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
semantic-version==2.8.5
six==1.14.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
termcolor==1.1.0
tzdata==2023.3
Unipath==1.1
urllib3==1.26.16
wcwidth==0.1.9
